Project re-structured. Common module created.
