+++ inteccl.py
@@ -1,13 +1,22 @@
 import click
 import modules.user_module as UM
+from bs4 import BeautifulSoup
+import modules.available_courses as AC
+import modules.support.core_authenticator as CA
 
 @click.command()
 @click.argument('command')
 @click.option('--id', '-i')
 @click.option('--password', '-p')
 def main(command, id, password):
+    br = CA.get_context(id, password)
+    raw_text = br.response().read()
+    context_page = BeautifulSoup(raw_text, 'html.parser')
+
     if command == 'user':
-        click.echo(UM.get_user(id, password))
+        click.echo(UM.get_user(context_page))
+    elif command == 'courses':
+        click.echo(AC.get_available_courses(context_page))
 
 if __name__ == "__main__":
     main()
