+++ modules/available_courses.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+import common as Common
+from bs4 import BeautifulSoup
+import json
+
+def get_available_courses(context_page):
+    data = {}
+
+    data['example'] = 'some text found...' 
+
+    return json.dumps(data)
